Problem 03. Inheritance, Problem 04. Interfaces
The main class is Inheritance. It creates two similar starter kits based on four stationery items, one an ArrayList - and then sorts it using Collections.Sort() (through the Stationery class implementing the Comparable interface) - and the other a TreeSet, which is instantiated with the use of a custom-built Comparator in the Stationery class, so the items are added in a sorted order.

The Stationery class serves as the superclass for two child classes - WritingInstruments and PaperProducts. The former is further the superclass for the Fineliner, Pen, and Pencil classes. The latter is the superclass for the Notebook, PhotoPaper, PlainPaper, and Stickers classes.