Problem 05. Generics, Enums
This application simulates a school curriculum. It enables the user to add courses to the curriculum from a pre-defined list (implemented using enums), enroll students in courses, grade all students in a course in one swoop, and display various listings such as of all courses in the curriculum, of students and their grades for a specific course, and of a student's grades across all courses.

The MainMenu class  is the main class that implements the main menu of the application through a mainMenu object that containes a switch-case construct, which calls appropriate methods based on user input, and through an infinite do-while(true) loop, which is broken when the user presses the key to exit the application. 

The CourseList enum encapsulates a list of available courses from which the user can add various courses to the school curriculum. Each enum constant has an id variable of type integer to allow for easy selection by the user by typing in the id instead of the name.

The School class encapsulates a school, which contains two arraylists - of all students and of all courses that have been added by the user. It features several methods, such as:
- the selectionFrom helper method, which takes an object of type ArrayList as the parameter that contains certain options from which the user has to make a selection, prints the options to the console, gets user input, validates it (not the InputMismatchException though, the input is presumed to be an integer, it only checks that the input is equal to one of the listed numbers) and returns it;
- the isCourseInCurriculum helper method, which takes an integer value (course number selected by the user) as the parameter and checks whether this course has already been added to the school curriculum;
- the addCourseToCurriculum method, which prints a listing of available courses (the CourseList enum), gets user input, validates it, creates a Course object and adds it to the listOfCourses arraylist variable of the School object;
- the showAvailableCourses method, which prints a list of all courses added to the school curriculum.

The Course class encapsulates a school course. A course can feature independently either integer or floating point grades. The Course class achieves this by introducing a type variable T that extends the Number class to Map<Student, T> studentAndGradeList that maps the students enrolled in the course to their grades. The class features the following methods:
- the fillStudents method, which fills students for a course. It firsly checks whether a student with the entered name has already been enrolled in the course (that is, exists in the Course object's studentAndGradeList). If not, the method then checks that in the school object's listOfStudents. If the student is found, it adds the course to the student's courseAndGradeList and in turn adds the student from the school object's listOfStudents to the course's studentAndGradeList. If there's no such student, the method creates a new Student objects with the entered first name / last name and puts it to the above lists;
- the gradeAllStudents method, which enables the user to grade all students enrolled in the course. It validates whether the grade type entered by the user (integer or floating point) corresponds to the grade type set at the creation of the course;
-the displayGrades method, which displays a listing of the students enrolled in the course along with their grades.

The Student class encapsulates a school student. Similarly to the Course class, it uses the same generic construct for Map<Course, T> courseAndGradeList. The class implements the Comparable interface and overrides the hashCode, compareTo and equals methods in order to be able to check two Student objects for logical equality (used by the Course class to check whether a student already exists). Aside from the above methods and the overriden toString method the class has a single displayGrades method that displays courses and grades for the invoking student.

Areas for future improvement:
- Add checks for InputMismatchExceptions for various inputs by the user.
- Add editing students' names or removing them from a course.
- Allow duplicate student names for different students by implementing student object ids.
- Grade courses only partly for specific students or for only those that haven't been graded yet.